+++ administrator/components/com_loginguard/tmpl/attempts/default.php
@@ -4,6 +4,7 @@ defined('_JEXEC') or die;
 
 use Joomla\CMS\HTML\HTMLHelper;
 use Joomla\CMS\Language\Text;
+use Joomla\CMS\Layout\LayoutHelper;
 use Joomla\CMS\Router\Route;
 
 HTMLHelper::_('behavior.core');
@@ -14,22 +15,22 @@ $listDirn  = $this->escape((string) $this->state->get('list.direction'));
 ?>
 <form action="<?php echo Route::_('index.php?option=com_loginguard&view=attempts'); ?>" method="post" name="adminForm" id="adminForm">
     <div id="j-main-container" class="j-main-container">
-        <?php echo HTMLHelper::_('searchtools.default', array('view' => $this)); ?>
+        <?php echo LayoutHelper::render('joomla.searchtools.default', ['view' => $this]); ?>
 
         <table class="table table-striped" id="loginguardAttemptsList">
             <caption class="visually-hidden"><?php echo Text::_('COM_LOGINGUARD_ATTEMPTS_TITLE'); ?></caption>
             <thead>
                 <tr>
-                    <th scope="col"><?php echo HTMLHelper::_('grid.sort', 'JGLOBAL_FIELD_ID_LABEL', 'id', $listDirn, $listOrder); ?></th>
-                    <th scope="col"><?php echo HTMLHelper::_('grid.sort', 'COM_LOGINGUARD_HEADING_IP_ADDRESS', 'ip_address', $listDirn, $listOrder); ?></th>
-                    <th scope="col"><?php echo HTMLHelper::_('grid.sort', 'COM_LOGINGUARD_HEADING_NAME', 'name', $listDirn, $listOrder); ?></th>
-                    <th scope="col"><?php echo HTMLHelper::_('grid.sort', 'COM_LOGINGUARD_HEADING_USERNAME', 'username', $listDirn, $listOrder); ?></th>
-                    <th scope="col"><?php echo HTMLHelper::_('grid.sort', 'COM_LOGINGUARD_HEADING_STATUS', 'status', $listDirn, $listOrder); ?></th>
-                    <th scope="col"><?php echo HTMLHelper::_('grid.sort', 'COM_LOGINGUARD_HEADING_DATETIME', 'created', $listDirn, $listOrder); ?></th>
-                    <th scope="col"><?php echo HTMLHelper::_('grid.sort', 'COM_LOGINGUARD_HEADING_COUNTRY', 'country', $listDirn, $listOrder); ?></th>
-                    <th scope="col"><?php echo HTMLHelper::_('grid.sort', 'COM_LOGINGUARD_HEADING_BROWSER', 'browser', $listDirn, $listOrder); ?></th>
-                    <th scope="col"><?php echo HTMLHelper::_('grid.sort', 'COM_LOGINGUARD_HEADING_OS', 'operating_system', $listDirn, $listOrder); ?></th>
-                    <th scope="col"><?php echo HTMLHelper::_('grid.sort', 'COM_LOGINGUARD_HEADING_WHERE', 'client', $listDirn, $listOrder); ?></th>
+                    <th scope="col"><?php echo HTMLHelper::_('searchtools.sort', 'JGLOBAL_FIELD_ID_LABEL', 'id', $listDirn, $listOrder); ?></th>
+                    <th scope="col"><?php echo HTMLHelper::_('searchtools.sort', 'COM_LOGINGUARD_HEADING_IP_ADDRESS', 'ip_address', $listDirn, $listOrder); ?></th>
+                    <th scope="col"><?php echo HTMLHelper::_('searchtools.sort', 'COM_LOGINGUARD_HEADING_NAME', 'name', $listDirn, $listOrder); ?></th>
+                    <th scope="col"><?php echo HTMLHelper::_('searchtools.sort', 'COM_LOGINGUARD_HEADING_USERNAME', 'username', $listDirn, $listOrder); ?></th>
+                    <th scope="col"><?php echo HTMLHelper::_('searchtools.sort', 'COM_LOGINGUARD_HEADING_STATUS', 'status', $listDirn, $listOrder); ?></th>
+                    <th scope="col"><?php echo HTMLHelper::_('searchtools.sort', 'COM_LOGINGUARD_HEADING_DATETIME', 'created', $listDirn, $listOrder); ?></th>
+                    <th scope="col"><?php echo HTMLHelper::_('searchtools.sort', 'COM_LOGINGUARD_HEADING_COUNTRY', 'country', $listDirn, $listOrder); ?></th>
+                    <th scope="col"><?php echo HTMLHelper::_('searchtools.sort', 'COM_LOGINGUARD_HEADING_BROWSER', 'browser', $listDirn, $listOrder); ?></th>
+                    <th scope="col"><?php echo HTMLHelper::_('searchtools.sort', 'COM_LOGINGUARD_HEADING_OS', 'operating_system', $listDirn, $listOrder); ?></th>
+                    <th scope="col"><?php echo HTMLHelper::_('searchtools.sort', 'COM_LOGINGUARD_HEADING_WHERE', 'where_at', $listDirn, $listOrder); ?></th>
                 </tr>
             </thead>
             <tbody>
@@ -49,7 +50,7 @@ $listDirn  = $this->escape((string) $this->state->get('list.direction'));
                             <td><?php echo $this->escape((string) $item->country); ?></td>
                             <td><?php echo $this->escape((string) $item->browser); ?></td>
                             <td><?php echo $this->escape((string) $item->operating_system); ?></td>
-                            <td><?php echo $this->escape((string) $item->client); ?></td>
+                            <td><?php echo $this->escape((string) ($item->where_at ?: $item->client)); ?></td>
                         </tr>
                     <?php endforeach; ?>
                 <?php endif; ?>